pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "feishu-claude"
version = "0.1.0"
description = "Feishu bot bridge to Claude Agent SDK"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.0.14",
    "lark-oapi>=1.4.0",
    "pyyaml>=6.0",
]

[project.scripts]
feishu-claude = "feishu_claude.main:main"

[tool.hatch.build.targets.wheel]
packages = ["feishu_claude"]
